requirements.txt
# SwarmRouter-v1 Requirements
# Python 3.11+ recommended

# Core ML
torch>=2.1.0
transformers>=4.36.0
datasets>=2.16.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0

# Training
trl>=0.7.0
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git

# API & CLI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
rich>=13.7.0

# Optional but recommended
# For GPU monitoring
# nvidia-ml-py3>=7.352.0

# For GGUF export (requires llama.cpp separately)
# numpy>=1.24.0
# sentencepiece>=0.1.99

# Dev/Testing
# pytest>=7.4.0
# black>=23.12.0
# ruff>=0.1.9
